// src/Promise/components/ArchiveSchema/interfaces.ts
export interface GinUploadParam {
  action: string;
  maxCount?: number;
}

export interface GinSelectParam {
  action: string;
  multiple?: boolean;
}

export interface GinTreeSelectParam {
  action: string;
  multiple?: boolean;
}

export interface GinCheckboxParam {
  action: string;
}

export interface GinRadioParam {
  action: string;
}
export interface GinRadioButtonParam {
  action: string;
}